+++ cmd/server/main.go
@@ -25,15 +25,99 @@
 package main
 
 import (
+	"fmt"
 	"os"
+	"path"
+	"path/filepath"
 
+	"github.com/urfave/cli"
+
+	"go.temporal.io/server/common/headers"
 	_ "go.temporal.io/server/common/persistence/sql/sqlplugin/mysql"      // needed to load mysql plugin
 	_ "go.temporal.io/server/common/persistence/sql/sqlplugin/postgresql" // needed to load postgresql plugin
+	"go.temporal.io/server/common/service/config"
 	"go.temporal.io/server/temporal"
 )
 
 // main entry point for the temporal server
 func main() {
-	app := temporal.BuildCLI()
+	app := buildCLI()
 	_ = app.Run(os.Args)
+}
+
+// buildCLI is the main entry point for the temporal server
+func buildCLI() *cli.App {
+	app := cli.NewApp()
+	app.Name = "temporal"
+	app.Usage = "Temporal server"
+	app.Version = headers.ServerVersion
+
+	app.Flags = []cli.Flag{
+		cli.StringFlag{
+			Name:   "root, r",
+			Value:  getDefaultRootDir(),
+			Usage:  "Root directory of execution environment",
+			EnvVar: config.EnvKeyRoot,
+		},
+		cli.StringFlag{
+			Name:   "config, c",
+			Value:  "config",
+			Usage:  "Config dir path relative to root",
+			EnvVar: config.EnvKeyConfigDir,
+		},
+		cli.StringFlag{
+			Name:   "env, e",
+			Value:  "development",
+			Usage:  "Runtime environment",
+			EnvVar: config.EnvKeyEnvironment,
+		},
+		cli.StringFlag{
+			Name:   "zone, az",
+			Usage:  "Availability zone",
+			EnvVar: config.EnvKeyAvailabilityZone,
+		},
+	}
+
+	allServicesStringSlice := cli.StringSlice(temporal.Services)
+	app.Commands = []cli.Command{
+		{
+			Name:  "start",
+			Usage: "Start Temporal server",
+			Flags: []cli.Flag{
+				cli.StringSliceFlag{
+					Name:  "services, s",
+					Value: &allServicesStringSlice,
+					Usage: "List of services to start",
+				},
+			},
+			Action: func(c *cli.Context) error {
+				env := c.GlobalString("env")
+				zone := c.GlobalString("zone")
+				configDir := path.Join(c.GlobalString("root"), c.GlobalString("config"))
+
+				services := c.StringSlice("services")
+
+				s := temporal.NewServer(
+					temporal.ForServices(services),
+					temporal.WithConfigLoader(configDir, env, zone),
+					temporal.InterruptOn(temporal.InterruptCh()),
+				)
+
+				err := s.Start()
+				if err != nil {
+					return cli.NewExitError(fmt.Sprintf("Unable to start server: %v.", err), 1)
+				}
+				return cli.NewExitError("All services are shutdown.", 0)
+			},
+		},
+	}
+	return app
+}
+
+func getDefaultRootDir() string {
+	exe, err := os.Executable()
+	if err != nil {
+		return "."
+	}
+	return filepath.Dir(exe)
 }
